Bundle fallback core module with Blender add-on
